Enhance permissions and validation across CRM modules
- Implemented permission checks for deleting activities, deals, queries, tasks, and returns to ensure only authorized users can perform these actions.
- Updated validation rules for deals, deliveries, invoices, and returns to enforce integer quantities and prevent invalid submissions.
- Improved user experience by conditionally displaying delete buttons based on user permissions in the UI.
- Added logic to prevent deletion of vendors that are still referenced in invoices or quotations to maintain data integrity.
- Enhanced refund calculations in the SalesReturn model to accurately reflect GST-inclusive amounts and ensure correct invoice updates.
- Refactored various components to improve code readability and maintainability.

+++ app/Http/Livewire/Crm/Deliveries/Create.php
@@ -61,7 +61,8 @@ class Create extends Component
                 preg_match('/^lines\.(\d+)\.delivery_qty$/', $attribute, $m);
                 $max = (float) ($this->lines[(int) $m[1]]['balance_qty'] ?? 0);
 
-                return ['required', 'numeric', 'min:0', 'max:'.$max];
+                // Whole units only — products are counted, not measured.
+                return ['required', 'integer', 'min:0', 'max:'.$max];
             }),
         ];
     }
